fix: use `pivot_root` instead of `chroot`
This commit fixes issues with `docker exec`. Changing the root with `chroot` does not actually change the root mount point but merely the way paths are resolved. This leads to issues with `docker exec` where the mount points are not set up correctly for the executed process. Using `pivot_root` instead resolves these issues.